+++ tests/Shutdown/PredisShutdownTest.php
@@ -4,50 +4,58 @@ declare(strict_types=1);
 namespace Tomaj\Hermes\Test\Shutdown;
 
 use PHPUnit\Framework\TestCase;
-use Tomaj\Hermes\Shutdown\RedisShutdown;
+use Predis\Client;
+use Predis\Response\Status;
+use Tomaj\Hermes\Shutdown\PredisShutdown;
 
 /**
  * Class RedisShutdownTest
  *
  * @package Tomaj\Hermes\Test\Shutdown
- * @covers \Tomaj\Hermes\Shutdown\RedisShutdown
+ * @covers \Tomaj\Hermes\Shutdown\PredisShutdown
  */
-class RedisShutdownTest extends TestCase
+class PredisShutdownTest extends TestCase
 {
-    public function testShouldShutdownWithoutRedisEntry()
+    public function testShouldShutdownWithoutPredisEntry()
     {
-        $redis = $this->createMock(\Redis::class);
+        $redis = $this->getMockBuilder(Client::class)
+            ->addMethods(['get'])
+            ->getMock();
         $redis->expects($this->once())
             ->method('get')
             ->willReturn(null);
 
-        $redisShutdown = new RedisShutdown($redis);
+        $redisShutdown = new PredisShutdown($redis);
         $this->assertFalse($redisShutdown->shouldShutdown(new \DateTime()));
     }
 
     public function testShouldShutdownWithFutureEntry()
     {
         $futureTime = (new \DateTime())->modify('+1 month')->format('U');
-        $redis = $this->createMock(\Redis::class);
+        $redis = $this->getMockBuilder(Client::class)
+            ->addMethods(['get'])
+            ->getMock();
         $redis->expects($this->once())
             ->method('get')
             ->with('hermes_shutdown')
             ->willReturn($futureTime);
 
-        $redisShutdown = new RedisShutdown($redis);
+        $redisShutdown = new PredisShutdown($redis);
         $this->assertFalse($redisShutdown->shouldShutdown(new \DateTime()));
     }
 
     public function testShouldShutdownWithEntryAfterStartTime()
     {
         $pastTime = (new \DateTime())->modify('-1 month')->format('U');
-        $redis = $this->createMock(\Redis::class);
+        $redis = $this->getMockBuilder(Client::class)
+            ->addMethods(['get'])
+            ->getMock();
         $redis->expects($this->once())
             ->method('get')
             ->with('hermes_shutdown')
             ->willReturn($pastTime);
 
-        $redisShutdown = new RedisShutdown($redis);
+        $redisShutdown = new PredisShutdown($redis);
         $this->assertFalse($redisShutdown->shouldShutdown(new \DateTime()));
     }
 
@@ -55,26 +63,46 @@ class RedisShutdownTest extends TestCase
     {
         $startTime = (new \DateTime())->modify('-1 hour');
         $shutdownTime = (new \DateTime())->modify('-5 minutes')->format('U');
-        $redis = $this->createMock(\Redis::class);
+        $redis = $this->getMockBuilder(Client::class)
+            ->addMethods(['get'])
+            ->getMock();
+
         $redis->expects($this->once())
             ->method('get')
             ->with('hermes_shutdown')
             ->willReturn($shutdownTime);
 
-        $redisShutdown = new RedisShutdown($redis);
+        $redisShutdown = new PredisShutdown($redis);
         $this->assertTrue($redisShutdown->shouldShutdown($startTime));
     }
 
     public function testShutdownStoredCorrectValueToRedis()
     {
         $shutdownTime = (new \DateTime())->modify('-5 minutes');
-        $redis = $this->createMock(\Redis::class);
+        $redis = $this->getMockBuilder(Client::class)
+            ->addMethods(['set'])
+            ->getMock();
         $redis->expects($this->once())
             ->method('set')
             ->with('hermes_shutdown', $shutdownTime->format('U'))
-            ->willReturn(true);
+            ->willReturn(new Status('OK'));
 
-        $redisShutdown = new RedisShutdown($redis);
+        $redisShutdown = new PredisShutdown($redis);
+        $this->assertTrue($redisShutdown->shutdown($shutdownTime));
+    }
+
+    public function testShutdownStoredCorrectValueToRedisPredis()
+    {
+        $shutdownTime = (new \DateTime())->modify('-5 minutes');
+        $redis = $this->getMockBuilder(\Predis\Client::class)
+            ->addMethods(['set'])
+            ->getMock();
+        $redis->expects($this->once())
+            ->method('set')
+            ->with('hermes_shutdown', $shutdownTime->format('U'))
+            ->willReturn(new Status('OK'));
+
+        $redisShutdown = new PredisShutdown($redis);
         $this->assertTrue($redisShutdown->shutdown($shutdownTime));
     }
 }